+++ Optimizer/UWPHelper.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Management.Automation;
 
 namespace Optimizer
@@ -8,6 +9,10 @@ namespace Optimizer
         internal static List<KeyValuePair<string, string>> GetUWPApps(bool showAll)
         {
             List<KeyValuePair<string, string>> modernApps = new List<KeyValuePair<string, string>>();
+            if (Utilities.CurrentWindowsVersion == WindowsVersion.Windows8)
+            {
+                showAll = true;
+            }
 
             using (PowerShell script = PowerShell.Create())
             {
@@ -21,8 +26,19 @@ namespace Optimizer
                 }
 
                 string[] tmp;
+                Collection<PSObject> psResult;
+                try
+                {
+                    psResult = script.Invoke();
+                }
+                catch
+                {
+                    return modernApps;
+                }
 
-                foreach (PSObject x in script.Invoke())
+                if (psResult == null) return modernApps;
+
+                foreach (PSObject x in psResult)
                 {
                     tmp = x.ToString().Replace("@", string.Empty).Replace("{", string.Empty).Replace("}", string.Empty).Replace("Name=", string.Empty).Replace("InstallLocation=", string.Empty).Trim().Split(';');
                     modernApps.Add(new KeyValuePair<string, string>(tmp[0], tmp[1]));